refactor: TravelPlan 패키지 분리
