+++ src/main/java/com/example/travelday/domain/travelplan/dto/request/batch/TravelPlanListReqDto.java
@@ -1,5 +1,6 @@
-package com.example.travelday.domain.travelroom.dto.request;
+package com.example.travelday.domain.travelplan.dto.request.batch;
 
+import com.example.travelday.domain.travelplan.dto.request.TravelPlanReqDto;
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Builder;
